pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doctr_process"
version = "0.1.0"
description = "OCR pipeline for tickets/invoices"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-doctr[torch]",
    "python-dotenv>=1.0",
    "pydantic>=2,<3",
    "loguru>=0.7",
    "pandas",
    "tqdm",
    "pytesseract",
    "pdf2image",
    "pypdf",
    "PyPDF2",
    "pymupdf",
    "opencv-python-headless",
    "pillow",
    "openpyxl",
    "pyyaml",
    "office365-rest-python-client",
    "easyocr",
    "requests",
    "importlib-resources>=1.3; python_version<'3.9'",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy"]

[project.scripts]
doctr-process = "doctr_process.main:main"
doctr-gui = "doctr_process.gui:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
doctr_process = ["configs/*", "assets/*", "assets/branding/*"]

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "T20"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
extend-exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
